% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{articulations}
\alias{articulations}
\title{Single note articulations and syntax}
\format{
A data frame with 3 column and 44 rows.
}
\usage{
articulations
}
\description{
A data frame containing categorized sets of articulations that can be used
in phrase construction.
}
\keyword{datasets}
