% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{as_music_df}
\alias{as_music_df}
\title{Noteworthy string to data frame}
\usage{
as_music_df(notes, key = NULL, scale = "diatonic", chords = c("root",
  "list", "character"), si_format = c("mmp_abb", "mmp", "ad_abb", "ad"))
}
\arguments{
\item{notes}{character, a noteworthy string.}

\item{key}{character, key signature, only required for inclusion of scale
degrees.}

\item{scale}{character, defaults to \code{"diatonic"}. Only used in
conjunction with \code{key}, this can be used to alter scale degrees. Not
any arbitrary combination of valid \code{key} and valid \code{scale} is
valid. See \code{\link{scale_degree}}.}

\item{chords}{character, how to structure columns containing multiple values
per chord/row of data frame. See details.}

\item{si_format}{character, format for scale intervals. See
\code{\link{scale_interval}}.}
}
\value{
a tibble data frame
}
\description{
Convert a noteworthy string to a tibble data frame and include additional
derivative variables.
}
\details{
For some derived column variables the root note (lowest pitch) in chord is
used. This is done for pitch intervals and scale intervals between adjacent
timesteps. This also occurs for scale degrees.

\code{chord = "root"} additionally collapses columns like semitone, octave,
and frequency to the value for the root note so that all rows contain one
numeric value. \code{chord = "list"} retains full information as list
columns. \code{chord = "character"} collapses into strings so that values are
readily visible when printing the table, but information is not stripped and
can be recovered without recomputing from the original pitches.
}
\examples{
x <- "a, b, c d e f g# a ac'e' a c' e' c' a"
as_music_df(x, key = "c", scale = "major")
as_music_df(x, key = "am", scale = "harmonic_minor", si_format = "ad_abb")
}
