% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chord_mapping.R
\name{chord-mapping}
\alias{chord-mapping}
\alias{gc_info}
\alias{gc_fretboard}
\alias{chord_is_known}
\alias{split_chord_name}
\alias{chord_name_root}
\alias{chord_name_mod}
\title{Chord mapping}
\usage{
gc_info(name, root_fret = NA, min_fret = NA, bass_string = NA,
  open = NA, key = "c", ignore_octave = FALSE)

gc_fretboard(name, root_fret = NA, min_fret = NA, bass_string = NA,
  open = NA, key = "c", ignore_octave = FALSE)

chord_is_known(notes)

split_chord_name(name)

chord_name_root(name)

chord_name_mod(name)
}
\arguments{
\item{name}{character, chord name in \code{tabr} format, e.g., \code{"bM b_m b_m7#5"}, etc.}

\item{root_fret}{integer, optional filter for chords whose root note matches a specific fret. May be a vector.}

\item{min_fret}{integer, optional filter for chords whose notes are all at or above a specific fret. May be a vector.}

\item{bass_string}{integer, optional filter for chords whose lowest pitch string matches a specific string, 6, 5, or 4. May be a vector.}

\item{open}{logical, optional filter for open and movable chords. \code{NA} retains both types.}

\item{key}{character, key signature, used to enforce type of accidentals.}

\item{ignore_octave}{logical, if \code{TRUE}, functions like \code{gc_info} and \code{gc_fretboard} return more results.}

\item{notes}{character, a noteworthy string.}
}
\value{
various, see details regarding each function.
}
\description{
Helper functions for chord mapping.
}
\details{
These functions assist with mapping between different information that define chords.

For \code{chord_is_known}, a check is done against chords in the \code{guitarChords} dataset.
A simple noteworthy string is permitted, but any single-note entry will automatically yield a \code{FALSE} result.

\code{gc_info} returns a tibble data frame containing complete information for the subset of predefined guitar chords specified by \code{name} and \code{key}.
Any accidentals present in the chord root of \code{name} (but not in the chord modifier, e.g., \code{m7_5} or \code{m7#5}) are converted according to \code{key} if necessary.

\code{split_chord_name} splits a vector or space-delimited set of chord names into a tibble data frame containing separate chord root and chord modifier columns.
\code{chord_name_root} and \code{chord_name_mod} are simple wrappers around this.
}
\examples{
chord_is_known("a b_,fb_d'f'")

chord_name_root("a aM b_,m7#5")
chord_name_mod("a aM b_,m7#5")

gc_info("a") # a major chord, not a single note
gc_info("ceg a#m7_5") # only third entry is a guitar chord
gc_info("ceg a#m7_5", key = "f")

gc_info("a,m c d f,")
gc_fretboard("a,m c d f,", 0:3)
}
