% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chords.R
\name{chord_break}
\alias{chord_break}
\title{Broken chords}
\usage{
chord_break(notes)
}
\arguments{
\item{notes}{character, noteworthy string that may contain chords.}
}
\value{
character
}
\description{
Convert chords in a noteworthy string or vector to broken chords.
}
\examples{
chord_break("c e g ceg ceg")
}
