% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guitarChords}
\alias{guitarChords}
\title{Predefined guitar chords.}
\format{A data frame with 13 columns.}
\usage{
guitarChords
}
\description{
A data frame containing information for many predefined guitar chords.
}
\keyword{datasets}
