% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phrase_helpers.R
\name{hp}
\alias{hp}
\title{Hammer ons and pull offs}
\usage{
hp(...)
}
\arguments{
\item{...}{character, note durations. Numeric is allowed for lists of single
inputs. See examples.}
}
\value{
character.
}
\description{
Helper function for generating hammer on and pull off syntax.
}
\details{
This is especially useful for repeated instances. This function applies to
general slur notation as well.
Multiple input formats are allowed. Total number of note durations must be
even because all slurs require start and stop points.
}
\examples{
hp(16, 16)
hp("16 16")
hp("16 8 16", "8 16 8")
}
