% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{interval_semitones}
\alias{interval_semitones}
\title{Interval semitones}
\usage{
interval_semitones(interval)
}
\arguments{
\item{interval}{character, interval ID. See details.}
}
\value{
integer
}
\description{
Convert named intervals to numbers of semitones.
For a complete list of valid interval names and abbreviations see
\code{\link{mainIntervals}}. \code{interval} may be a vector.
}
\examples{
x <- c("minor third", "m3", "augmented second", "A2")
y <- c("P1", "m2", "M2", "m3", "M3", "P4", "TT", "P5")
interval_semitones(x)
interval_semitones(y)
}
\seealso{
\code{\link{mainIntervals}}
}
