% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{lilypond}
\alias{lilypond}
\title{Save score to LilyPond file}
\usage{
lilypond(score, file, key = "c", time = "4/4", tempo = "2 = 60",
  header = NULL, paper = NULL, string_names = NULL, endbar = TRUE,
  midi = TRUE, path = NULL, crop_png = TRUE)
}
\arguments{
\item{score}{a score object.}

\item{file}{character, LilyPond output file ending in \code{.ly}. May
include an absolute or relative path.}

\item{key}{character, key signature, e.g., \code{c}, \code{b_}, \code{f#m},
etc.}

\item{time}{character, defaults to \code{"4/4"}.}

\item{tempo}{character, defaults to \code{"2 = 60"}.}

\item{header}{a named list of arguments passed to the header of the
LilyPond file. See details.}

\item{paper}{a named list of arguments for the LilyPond file page layout.
See details.}

\item{string_names}{label strings at beginning of tab staff. \code{NULL}
(default) for non-standard tunings only, \code{TRUE} or \code{FALSE} for
force on or off completely.}

\item{endbar}{character, the end bar.}

\item{midi}{logical, add midi inclusion specification to LilyPond file.}

\item{path}{character, optional output directory prefixed to \code{file},
may be an absolute or relative path. If \code{NULL} (default), only
\code{file} is used.}

\item{crop_png}{logical, alter template for cropped height. See
details.}
}
\value{
nothing returned; a file is written.
}
\description{
Write a score to a LilyPond format (\code{.ly}) text file for later use by
LilyPond or subsequent editing outside of R.
}
\details{
All \code{header} list elements are character strings. The options for
\code{header} include:
\itemize{
  \item \code{title}
  \item \code{subtitle}
  \item \code{composer}
  \item \code{album}
  \item \code{arranger}
  \item \code{instrument}
  \item \code{meter}
  \item \code{opus}
  \item \code{piece}
  \item \code{poet}
  \item \code{copyright}
  \item \code{tagline}
}

All \code{paper} list elements are numeric except \code{page_numbers} and
\code{print_first_page_number},
which are logical. \code{page_numbers = FALSE} suppresses all page numbering.
When \code{page_numbers = TRUE}, you can set
\code{print_first_page_number = FALSE} to suppress printing of only the
first page number. \code{first_page_number} is the number of the first page,
defaulting to 1, and determines all subsequent page numbers. These arguments
correspond to LilyPond paper block variables.

The options for \code{paper} include the following and have the following
default values if not provided:
\itemize{
  \item \code{textheight = 220}
  \item \code{linewidth = 150}
  \item \code{indent = 0}
  \item \code{fontsize = 10}
  \item \code{page_numbers = TRUE}
  \item \code{print_first_page_number = TRUE}
  \item \code{first_page_number = 1}
}

By default \code{crop_png = TRUE}. This alters the template so that when
the LilyPond output file is created, it contains specifications for cropping
the image to the content when that file is rendered by LilyPond to png.
The image will have its width and height automatically cropped
rather than retain the standard page dimensions.
This only applies to png outputs made from the LilyPond file, not pdf.
The argument is also ignored if explicitly providing \code{textheight} to
\code{paper}. You may still provide \code{linewidth} to \code{paper} if you
find you need to increase it beyond the default 150mm, generally as a result
of using a large \code{fontsize}.
Various \code{render_*} functions that wrap \code{lilypond} make use of this
argument as well.
}
\examples{
x <- phrase("c ec'g' ec'g'", "4 4 2", "5 432 432")
x <- track(x)
x <- score(x)
outfile <- file.path(tempdir(), "out.ly")
lilypond(x, outfile)
}
\seealso{
\code{\link{tab}}, \code{\link{render_chordchart}},
\code{\link{midily}}
}
