% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mainIntervals}
\alias{mainIntervals}
\title{Main musical intervals}
\format{
A data frame with 5 columns and 26 rows
}
\usage{
mainIntervals
}
\description{
A data frame containing descriptions of the main intervals, simple and
compound.
}
\keyword{datasets}
