% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music.R
\name{music}
\alias{music}
\alias{musical}
\alias{as_music}
\alias{is_music}
\alias{music_split}
\alias{music_notes}
\alias{music_info}
\alias{music_tsig}
\title{Check music string validity}
\usage{
musical(x)

as_music(notes, info = NULL, accidentals = NULL, tsig = "4/4",
  format = NULL)

is_music(x)

music_split(x)

music_notes(x)

music_info(x)

music_tsig(x)
}
\arguments{
\item{x}{character, a music string. May or may not have the 'music' class.}

\item{notes, info}{noteworthy and note info strings. For \code{as_music}, a
complete music string is assumed for \code{notes} when \code{info = NULL}.}

\item{accidentals}{\code{NULL} or character, represent accidentals,
\code{"flat"} or \code{"sharp"}.}

\item{tsig}{character, store the time signature as a music attribute.
Defaults to \code{"4/4"}.}

\item{format}{\code{NULL} or character, the timestep delimiter format,
\code{"space"} or \code{"vector"}.}
}
\value{
depends on the function
}
\description{
Check whether a string is comprised exclusively of valid syntax for music
strings. A music object can be built from such a string. It combines a
noteworthy string and a note info string.
}
\details{
\code{musical} returns a scalar logical result indicating whether all
timesteps contain exclusively valid entries.

\code{as_music} can be used to coerce to the \code{music} class.
Coercion will fail if the string is not musical.
The \code{music} class has its own \code{print} and \code{summary} methods.

For \code{as_music} and in general for functions that accept
When \code{accidentals}, \code{format}, and \code{tsig} are \code{NULL},
these settings are inferred from the musical string input.
When mixed formats are present, flats are the default for accidentals and
4/4 common time is the default time signature.
}
\examples{
x <- "a#4] b_4.. c,4x d''t8) et8)( g_'t8)- a4 c,e_,g,4 ce_g1"
is_music(x)
musical(x)
x <- as_music(x)
is_music(x)
x

summary(x)

attributes(x)

music_split(x)
music_notes(x)
music_info(x)
music_tsig(x)

as_music(x, accidentals = "sharp")
}
\seealso{
\code{\link{note-checks}}, \code{\link{note-metadata}},
\code{\link{note-summaries}}, \code{\link{note-coerce}}
}
