% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes.R
\name{note-metadata}
\alias{note-metadata}
\alias{n_steps}
\alias{n_notes}
\alias{n_chords}
\alias{n_octaves}
\alias{tally_notes}
\alias{tally_pitches}
\alias{tally_octaves}
\alias{distinct_notes}
\alias{distinct_pitches}
\alias{distinct_octaves}
\alias{pitch_range}
\alias{semitone_range}
\alias{octave_range}
\alias{octave_type}
\alias{accidental_type}
\alias{time_format}
\title{Noteworthy string metadata}
\usage{
n_steps(notes)

n_notes(notes)

n_chords(notes)

n_octaves(notes)

tally_notes(notes)

tally_pitches(notes)

tally_octaves(notes)

distinct_notes(notes)

distinct_pitches(notes)

distinct_octaves(notes)

pitch_range(notes)

semitone_range(notes)

octave_range(notes)

octave_type(notes)

accidental_type(notes)

time_format(notes)
}
\arguments{
\item{notes}{character, a noteworthy string, space-delimited or vector of individual entries.}
}
\value{
integer or character
}
\description{
Inspect metadata for noteworthy strings.
}
\examples{
x <- "e_2 a_, c#f#a#"
n_steps(x)
n_notes(x)
n_chords(x)
n_octaves(x)

tally_notes(x)
tally_pitches(x)
tally_octaves(x)
distinct_notes(x)
distinct_pitches(x)
distinct_octaves(x)

pitch_range(x)
semitone_range(x)
octave_range(x)

octave_type(x)
accidental_type(x)
time_format(x)
}
