% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noteinfo.R
\name{noteinfo}
\alias{noteinfo}
\alias{phrase_duration}
\title{Note info helpers}
\usage{
phrase_duration(phrase)
}
\arguments{
\item{phrase}{a phrase object.}
}
\value{
character
}
\description{
Functions for working with note info strings.
}
\examples{
a <- notate("8x", "Start here")
notes <- "a, b, c d e f g# a r ac'e' a c' e' c' r*3 ac'e'~ ac'e'"
info <- paste(a, "8] 8] 16 4.. 16- 16 2 2 4. 8( 4)( 4) 8*4 1 1")
p1 <- phrase(notes, info)

phrase_duration(p1)
}
