% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{pitch_freq}
\alias{pitch_freq}
\alias{freq_pitch}
\title{Pitch-frequency conversion}
\usage{
pitch_freq(notes, fixed_note = "a4", fixed_freq = 440)

freq_pitch(freq, fixed_note = "a4", fixed_freq = 440,
  collapse = FALSE, ...)
}
\arguments{
\item{notes}{character, note string, space-delimited or vector of individual entries.}

\item{fixed_note}{the fixed note, typically A4, the A above middle C.}

\item{fixed_freq}{the fixed frequency of \code{fixed_note}, typically 440 Hz.}

\item{freq}{numeric, frequencies in Hz.}

\item{collapse}{logical, collapse result into a single string.}

\item{...}{other arguments passed to \code{tranpose}, specifically \code{key} and \code{style}.}
}
\value{
numeric vector
}
\description{
Convert between pitch and frequency.
}
\details{
Frequencies are in Hertz. Values are based on the 12-tone equal-tempered scale. When converting an arbitrary frequency to pitch, it is rounded to the nearest pitch.
}
\examples{
x <- "a e4 a4 e5 a5"
y <- pitch_freq(x)
y

freq_pitch(y)

identical(as_noteworthy(x), freq_pitch(y, collapse = TRUE))
}
