% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{ratio_to_cents}
\alias{ratio_to_cents}
\alias{cents_to_ratio}
\title{Convert between chord frequency ratios and cents}
\usage{
ratio_to_cents(x, y = NULL)

cents_to_ratio(x)
}
\arguments{
\item{x}{a vector of ratios if \code{y = NULL}, otherwise frequencies. Cents for
\code{cents_to_ratio()}.}

\item{y}{if not \code{NULL}, frequencies and the ratios are given by \code{y / x}.}
}
\value{
numeric
}
\description{
Convert between frequency ratios and logarithmic cents
}
\examples{
ratio_to_cents(c(0.5, 1, 1.5, 2))
cents_to_ratio(c(-1200, 0, 701.955, 1200))
}
