% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midi.R
\name{read_midi}
\alias{read_midi}
\alias{midi_metadata}
\alias{midi_notes}
\alias{midi_time}
\alias{midi_key}
\title{Read and inspect MIDI file contents}
\usage{
read_midi(file)

midi_metadata(x)

midi_notes(x, channel = NULL, track = NULL)

midi_time(x)

midi_key(x)
}
\arguments{
\item{file}{character, path to MIDI file.}

\item{x}{a data frame returned by \code{read_midi}.}

\item{channel, track}{integer, filter rows on channel or track.}
}
\value{
a tibble data frame
}
\description{
Read MIDI file into a data frame and inspect the music data with supporting
functions.
}
\details{
This function wraps around \code{tuneR::readMidi} by Uwe Ligges and Johanna
Mielke.
}
\examples{
file <- system.file("example.mid", package = "tabr")
if(require("tuneR")){
  x <- read_midi(file)
  midi_metadata(x)
  midi_time(x)
  midi_key(x)
  midi_notes(x, channel = 1)
}
}
