% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{render_chordchart}
\alias{render_chordchart}
\title{Render a chord chart with LilyPond}
\usage{
render_chordchart(chords, file, keep_ly = FALSE, fontsize = 60,
  details = FALSE)
}
\arguments{
\item{chords}{named character vector of valid formatting for LilyPond chord
names and values. See examples.}

\item{file}{output file.}

\item{keep_ly}{logical, keep intermediate LilyPond file.}

\item{fontsize}{integer, vary this depending on the number of chords.
It displays properly if you keep the chords to a single page. If the full
set of chords does not fit then chords will be dropped from the rendered
output.}

\item{details}{logical, set to \code{FALSE} to disable printing of log
output to console.}
}
\value{
writes files to disk
}
\description{
Render a standalone chord chart of chord fretboard diagrams with LilyPond
for a set of chords.
}
\details{
This function uses a generates a LilyPond template for displaying only a
fretboard diagram chart. It then passes the file to LilyPond for rendering.
To plot specific fretboard diagrams in R using ggplot and with greater
control, use \code{fretboard_plot}.
}
\examples{
suppressPackageStartupMessages(library(dplyr))

chords <- filter(
  guitarChords, root \%in\% c("c", "f") & id \%in\% c("7", "M7", "m7") &
  !grepl("#", notes) & root_fret <= 12) \%>\%
  arrange(root, id)
chords <- setNames(chords$fretboard, chords$lp_name)
head(chords)

# requires LilyPond installation
if(tabr_options()$lilypond != ""){
  outfile <- file.path(tempdir(), "out.pdf")
  render_chordchart(chords, outfile, fontsize = 30)
}
}
