% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rest}
\alias{rest}
\title{Create rests}
\usage{
rest(x, n = 1)
}
\arguments{
\item{x}{character, LilyPond rest syntax.}

\item{n}{integer, number of repetitions.}
}
\value{
a character string.
}
\description{
Create multiple rests efficiently.
}
\examples{
rest("r1 r8", c(1, 4))
}
