% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale-helpers}
\alias{scale-helpers}
\alias{scale_diatonic}
\alias{scale_major}
\alias{scale_minor}
\alias{scale_chromatic}
\title{Scale helpers}
\usage{
scale_diatonic(key = "c", collapse = FALSE)

scale_major(key = "c", collapse = FALSE)

scale_minor(key = "am", collapse = FALSE)

scale_chromatic(root = "c", collapse = FALSE, sharp = TRUE)
}
\arguments{
\item{key}{character, key signature.}

\item{collapse}{logical, collapse result into a single string ready for phrase construction.}

\item{root}{character, root note.}

\item{sharp}{logical, accidentals in arbitrary scale output should be sharp rather than flat.}
}
\value{
character
}
\description{
Helper functions for working with musical scales.
}
\details{
For valid key signatures, see \code{\link{keys}}.
}
\examples{
scale_major("d")
scale_minor("dm")
scale_chromatic("a")
}
\seealso{
\code{\link{keys}}, \code{\link{mode-helpers}}
}
