% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_chords}
\alias{scale_chords}
\title{Diatonic chords}
\usage{
scale_chords(root = "c", scale = "major", type = c("triad",
  "seventh"), collapse = FALSE, style = "default")
}
\arguments{
\item{root}{character, root note or starting position of scale.}

\item{scale}{character, a valid named scale, referring to one of the existing \code{scale_*} functions.}

\item{type}{character, type of chord, either triad or seventh.}

\item{collapse}{logical, collapse result into a single string ready for phrase construction.}

\item{style}{character, passed to \code{transpose}.}
}
\value{
character
}
\description{
Obtain an ordered sequence of the diatonic chords for a given scale.
}
\examples{
scale_chords("d", "major")
scale_chords("d", "minor")
scale_chords("d", "harmonic minor")
# scale_chords("d", "melodic minor") TODO
}
