% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{sf_phrase}
\alias{sf_phrase}
\alias{sfp}
\title{Create a musical phrase from string/fret combinations}
\usage{
sf_phrase(string, fret, info, tuning = "standard", key = "c",
  style = c("tick", "integer"), bar = FALSE)

sfp(...)
}
\arguments{
\item{string}{character, string numbers associated with notes.}

\item{fret}{character, fret numbers associated with notes.}

\item{info}{character, metadata associated with notes.}

\item{tuning}{character, instrument tuning.}

\item{key}{character, key signature. Passed to \code{transpose} function.}

\item{style}{character, octave numbering style, tick or integer.}

\item{bar}{logical, insert a bar check at the end of the phrase.}

\item{...}{arguments passed to \code{sf_phrase}.}
}
\value{
a phrase.
}
\description{
Create a musical phrase from character strings that define string numbers, fret numbers and note metadata. This function is a wrapper around \code{\link{phrase}}.
It allows for specifying string/fret combinations instead of unambiguous pitch as is used by \code{phrase}.
In order to remove ambiguity, it is critical to specify the instrument string tuning and key signature.
See the main function \code{phrase} for more details.
}
\details{
The function \code{sfp} is a convenient shorthand wrapper for \code{sf_phrase}.
}
\examples{
sf_phrase("5 4 3 2 1", "1 3 3 3 1", "8*4 1", key = "b_")
sf_phrase("654321 6s 12 1 21", "133211 355333 11 (13) (13)(13)", "4 4 8 8 4", key = "f")
sfp("6s*2 1*4", "000232*2 2*4", "4 4 8*4", tuning = "dropD", key = "d")
}
\seealso{
\code{\link{phrase}}
}
