% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{tabr-rev}
\alias{tabr-rev}
\alias{rev.noteworthy}
\alias{rev.noteinfo}
\alias{rev.music}
\alias{rev.lyrics}
\title{Reverse for tabr classes}
\usage{
\method{rev}{noteworthy}(x)

\method{rev}{noteinfo}(x)

\method{rev}{music}(x)

\method{rev}{lyrics}(x)
}
\arguments{
\item{x}{object.}
}
\description{
Several methods are implemented for the classes \code{noteworthy}, \code{noteinfo}, and
\code{music}. See \code{\link[=tabr-methods]{tabr-methods()}} for more details on methods for tabr classes.
}
\examples{
# noteworthy class examples
x <- "a b c"
rev(x)
rev(as_noteworthy(x))

# noteinfo class examples
x <- "4x 4-.*2 2"
rev(as_noteinfo(x))

# music class examples
x <- "c,~4 c,1 c'e_'g'4-."
rev(as_music(x))
}
\seealso{
\code{\link[=tabr-methods]{tabr-methods()}}, \code{\link[=note-metadata]{note-metadata()}}
}
