% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{tabr_dep}
\alias{tabr_dep}
\alias{dup}
\alias{glue}
\title{Deprecated}
\usage{
dup(x, n = 1)

glue(...)
}
\arguments{
\item{x}{character, phrase or non-phrase string.}

\item{n}{integer, number of repetitions.}

\item{...}{character, phrase or non-phrase string.}
}
\description{
Deprecated tabr functions.
}
