% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{tabr_options}
\alias{tabr_options}
\title{Options}
\usage{
tabr_options(...)
}
\arguments{
\item{...}{a list of options.}
}
\value{
The function prints all set options if called with no arguments. When setting options, nothing is returned.
}
\description{
Options for tabr package.
}
\details{
Currently only \code{lilypond} is used. If the system path for the LilyPond executable is not stored in the system path environmental variables, it must be provided by the user after loading the package.
}
\examples{
tabr_options()
tabr_options(lilypond = "C:/Program Files (x86)/LilyPond/usr/bin/lilypond.exe")
}
