% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tabrSyntax}
\alias{tabrSyntax}
\title{tabr syntax}
\format{
A data frame with 3 columns for syntax description, operators and
examples.
}
\usage{
tabrSyntax
}
\description{
A data frame containing descriptions of syntax used in phrase construction
in tabr.
}
\keyword{datasets}
