% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phrase_helpers.R
\name{tie}
\alias{tie}
\alias{untie}
\title{Tied notes}
\usage{
tie(x)

untie(x)
}
\arguments{
\item{x}{character, a single chord.}
}
\value{
a character string.
}
\description{
Tie notes efficiently.
}
\details{
This function is useful for bar chords.
}
\examples{
tie("e,b,egbe'")
}
