% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{track}
\alias{track}
\title{Create a music track}
\usage{
track(phrase, add_staff = NA, voice = 1L)
}
\arguments{
\item{phrase}{a phrase object.}

\item{add_staff}{add a standard sheet music staff above the tablature staff. See details.}

\item{voice}{integer, ID indicating phrases belonging to unique voices within a single track. May be a vector.}
}
\value{
a track table.
}
\description{
Create a music track from a collection of musical phrases.
}
\details{
Musical phrases generated by \code{\link{phrase}} are fortified in a track table.
The default for an additional staff is \code{add_staff = "treble_8"}, which is commonly displayed in quality guitar tablature above the tablature staff to include precise rhythm and timing information.
Setting this to \code{NA} will not add another staff, which is appropriate for simple tracks where the additional space required would be wasteful.
Note that guitar is a transposing instrument.
}
\examples{
x <- phrase("c ec'g' ec'g'", "4 4 2", "5 432 432")
track(x)
}
