% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{trackbind}
\alias{trackbind}
\title{Bind track tables}
\usage{
trackbind(...)
}
\arguments{
\item{...}{track tables.}
}
\value{
a track table.
}
\description{
Bind together track tables by row.
}
\details{
This function appends multiple track tables into a single track table for preparation of generating a multi-track score.
Regardless of whether a \code{tabstaff} ID column is present in the input tracks, one is added/overidden ensuring a unique integer ID for each track.
This is used to separate music staves in the sheet music/tablature output.
The actual ID values assigned to each track do not matter; only the order in which tracks are bound, first to last.
}
\examples{
x <- phrase("c ec'g' ec'g'", "4 4 2", "5 432 432")
trackbind(track(x), track(x))
}
