% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.R
\name{transpose}
\alias{transpose}
\alias{tp}
\title{Transpose pitch}
\usage{
transpose(x, n = 0, key = NULL)

tp(...)
}
\arguments{
\item{x}{character, a valid string of notes, the type passed to \code{phrase}.}

\item{n}{integer, positive or negative number of semitones to transpose.}

\item{key}{character, the new key signature after transposing \code{x}. If provided, this helps ensure proper use of sharps vs. flats.}

\item{...}{arguments passed to transpose.}
}
\value{
a character string.
}
\description{
Transpose pitch by a number of semitones.
}
\details{
This function transposes the pitch of notes in a valid character string. The string must be of the form passed to the \code{info} argument to \code{\link{phrase}}.

Transposing is not done on a phrase object. The notes in a phrase object have already been transformed to LilyPond syntax and mixed with other potentially complex and variable information.
Transposing is intended to be done on a string of notes prior to passing it to \code{phrase}. It will work on strings that use either integer or tick mark octave numbering formats.
The transposed result will be a string with integer octave numbering.
}
\examples{
transpose("a_3 b_4 c5", 0)
tp("a_3 b_4 c5", -1)
tp("a_3 b_4 c5", 1)
tp("a#3 b4 c#5", 11)
tp("a#3 b4 c#5", 12)
tp("a#3 b4 c#5", 13)
tp("a3 b4 c5", 2, key = "f")
tp("a3 b4 c5", 2, key = "g")
tp("a b' c''", 2, key = "f")
tp("a, b c'", 2, key = "g")
}
