% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunings.R
\docType{data}
\name{tunings}
\alias{tunings}
\title{Predefined instrument tunings.}
\format{A data frame with 2 columns for the tuning ID and corresponding pitches and 32 rows for all predefined tunings.}
\usage{
tunings
}
\description{
A data frame containing some predefined instrument tunings commonly used for guitar, bass, mandolin, banjo, ukulele and orchestral instruments.
}
\keyword{datasets}
