% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tuplet}
\alias{tuplet}
\title{Tuplets}
\usage{
tuplet(notes, n)
}
\arguments{
\item{notes}{character, notes.}

\item{n}{integer, tuplet beat, e.g., 8 for 8th note tuplet.}
}
\value{
character.
}
\description{
Helper function for generating tuplet syntax.
}
\details{
This is an initial version.
}
\examples{
tuplet("c c# d", 8)
}
