% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phrase.R
\name{volta}
\alias{volta}
\alias{pct}
\title{Create a volta/repeat phrase}
\usage{
volta(phrase, n = 1, endings = NULL, silent = FALSE)

pct(phrase, n = 1)
}
\arguments{
\item{phrase}{a phrase or basic string to be repeated.}

\item{n}{integer, number of repeats of \code{phrase} (one less than the total number of plays).}

\item{endings}{list of phrases or basic strings, alternate endings.}

\item{silent}{if \code{TRUE}, no text will be printed above the staff to indicate the number of repeats when greater than one and no \code{endings = NULL}. This is useful when another staff already displays the text.}
}
\value{
a phrase.
}
\description{
Create a repeat section in LilyPond readable format.
}
\details{
This function takes a musical phrase or even a basic string and converts it into a volta phrase for the LilyPond volta engraver.

Note that basic strings should still be interpretable as a valid musical phrase by LilyPond.
For example, a one-measure rest, \code{"r1"} does not need to be wrapped in a phrase class to work, nor would any other string explicitly written out in valid LilyPond syntax.
As always, see the LilyPond documentation if you are not familiar with LilyPond syntax.
}
\examples{
x <- phrase("c ec'g' ec'g'", "4 4 2", "5 432 432")
e1 <- phrase("a", 1, 5)
e2 <- phrase("b", 1, 5)
volta(x) # one repeat or 2 plays
volta(x, list(e1, e2)) # one repeat with alternate ending
volta(x, 4, list(e1, e2)) # multiple repeats but with only one alternate ending
}
