% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{chord_set}
\alias{chord_set}
\title{Generate a chord set}
\usage{
chord_set(x, id)
}
\arguments{
\item{x}{character, six-string chord description from lowest to highest pitch, strings 6 through 1, e.g., the open A minor chord in standard tuning is \code{"xo221o"}.}

\item{id}{character, the name of the chord in LilyPond readable format, e.g., \code{"a:m"}.}
}
\value{
a named list.
}
\description{
Generate a chord set for a music score.
}
\details{
The chord set is a named list. The names are the chords and the list elements are strings defining string and fret fingering readable by LilyPond.
Multiple chord positions can be defined for the same chord name.
When defining chords, you may also wish to define rests or silence for chords added to a score for placement above the staff in time,
where no chord is to be played, though this will be excluded from the chord set as unnecessary.
The chord set list output by \code{chord_set} is only used for top center placement of a full set of chord diagrams for a music score.
}
\examples{
chord_names <- c("e:m", "c", "d", "e:m", "d", "r", "s")
chord_positions <- c("xx997x", "x5553x", "x7775x", "ooo22o", "232oxx", NA, NA)
chords <- chord_set(chord_positions, chord_names)
}
