% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chords.R
\name{chords}
\alias{chords}
\alias{chord_min}
\alias{chord_maj}
\alias{chord_min7}
\alias{chord_dom7}
\alias{chord_7s5}
\alias{chord_maj7}
\alias{chord_min6}
\alias{chord_maj6}
\alias{chord_dim}
\alias{chord_dim7}
\alias{chord_m7b5}
\alias{chord_aug}
\alias{chord_5}
\alias{chord_sus2}
\alias{chord_sus4}
\alias{chord_dom9}
\alias{chord_7s9}
\alias{chord_maj9}
\alias{chord_add9}
\alias{chord_min9}
\alias{chord_madd9}
\alias{chord_min11}
\alias{chord_7s11}
\alias{chord_maj7s11}
\alias{chord_11}
\alias{chord_maj11}
\alias{chord_13}
\alias{chord_min13}
\alias{chord_maj13}
\alias{xm}
\alias{xM}
\alias{xm7}
\alias{x7}
\alias{x7s5}
\alias{xM7}
\alias{xm6}
\alias{xM6}
\alias{xdim}
\alias{xdim7}
\alias{xm7b5}
\alias{xaug}
\alias{x5}
\alias{xs2}
\alias{xs4}
\alias{x9}
\alias{x7s9}
\alias{xM9}
\alias{xadd9}
\alias{xm9}
\alias{xma9}
\alias{xm11}
\alias{x7s11}
\alias{xM7s11}
\alias{x_11}
\alias{xM11}
\alias{x_13}
\alias{xm13}
\alias{xM13}
\title{Chord constructors}
\usage{
chord_min(note, key = "c", ...)

chord_maj(note, key = "c", ...)

chord_min7(note, key = "c", ...)

chord_dom7(note, key = "c", ...)

chord_7s5(note, key = "c", ...)

chord_maj7(note, key = "c", ...)

chord_min6(note, key = "c", ...)

chord_maj6(note, key = "c", ...)

chord_dim(note, key = "c", ...)

chord_dim7(note, key = "c", ...)

chord_m7b5(note, key = "c", ...)

chord_aug(note, key = "c", ...)

chord_5(note, key = "c", ...)

chord_sus2(note, key = "c", ...)

chord_sus4(note, key = "c", ...)

chord_dom9(note, key = "c", ...)

chord_7s9(note, key = "c", ...)

chord_maj9(note, key = "c", ...)

chord_add9(note, key = "c", ...)

chord_min9(note, key = "c", ...)

chord_madd9(note, key = "c", ...)

chord_min11(note, key = "c", ...)

chord_7s11(note, key = "c", ...)

chord_maj7s11(note, key = "c", ...)

chord_11(note, key = "c", ...)

chord_maj11(note, key = "c", ...)

chord_13(note, key = "c", ...)

chord_min13(note, key = "c", ...)

chord_maj13(note, key = "c", ...)

xm(note, key = "c", ...)

xM(note, key = "c", ...)

xm7(note, key = "c", ...)

x7(note, key = "c", ...)

x7s5(note, key = "c", ...)

xM7(note, key = "c", ...)

xm6(note, key = "c", ...)

xM6(note, key = "c", ...)

xdim(note, key = "c", ...)

xdim7(note, key = "c", ...)

xm7b5(note, key = "c", ...)

xaug(note, key = "c", ...)

x5(note, key = "c", ...)

xs2(note, key = "c", ...)

xs4(note, key = "c", ...)

x9(note, key = "c", ...)

x7s9(note, key = "c", ...)

xM9(note, key = "c", ...)

xadd9(note, key = "c", ...)

xm9(note, key = "c", ...)

xma9(note, key = "c", ...)

xm11(note, key = "c", ...)

x7s11(note, key = "c", ...)

xM7s11(note, key = "c", ...)

x_11(note, key = "c", ...)

xM11(note, key = "c", ...)

x_13(note, key = "c", ...)

xm13(note, key = "c", ...)

xM13(note, key = "c", ...)
}
\arguments{
\item{note}{character, chord root note.}

\item{key}{key signature. See details.}

\item{...}{additional arguments to \code{transpose}. Only accepts \code{style}.}
}
\value{
character
}
\description{
These functions construct basic chord string notation from a root \code{note}.
}
\details{
Providing a \code{key} signature is used only to ensure flats or sharps for accidentals.
An additional set of aliases with efficient names, of the form \code{x*} where \code{*} is a chord modifier abbreviation, is provided to complement the set of \code{chord_*} functions.

These functions create standard chords, not the multi-octave spanning types of chords commonly played on guitar.
}
\examples{
chord_min("d")
chord_maj("d")
xM("d")
}
\seealso{
\code{\link{transpose}}
}
