% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{lilypond}
\alias{lilypond}
\title{Save score to LilyPond file}
\usage{
lilypond(score, file, key = "c", time = "4/4", tempo = "2 = 60",
  header = NULL, string_names = NULL, paper = NULL, endbar = TRUE,
  midi = TRUE, out_dir = ".")
}
\arguments{
\item{score}{a score object.}

\item{file}{LilyPond output file, should end in \code{.ly}.}

\item{key}{character, key signature, e.g., \code{c}, \code{b_}, \code{f#m}, etc.}

\item{time}{character, defaults to \code{"4/4"}.}

\item{tempo}{character, defaults to \code{"2 = 60"}.}

\item{header}{a named list of arguments passed to the header of the LilyPond file. See details.}

\item{string_names}{label strings at beginning of tab staff. \code{NULL} (default) for non-standard tunings only, \code{TRUE} or \code{FALSE} for force on or off completely.}

\item{paper}{a named list of arguments for the LilyPond file page layout. See details.}

\item{endbar}{character, the end bar.}

\item{midi}{logical, add midi inclusion specification to LilyPond file.}

\item{out_dir}{character, output directory, defaults to working directory.}
}
\value{
nothing returned; a file is written.
}
\description{
Write score to a LilyPond format (\code{.ly}) text file for later use by LilyPond or subsequent editing outside of R.
}
\details{
The options for \code{header} include \code{title}, \code{subtitle}, \code{composer}, \code{album}, \code{arranger}, \code{instrument}, \code{meter}, \code{opus}, \code{piece}, \code{poet}, \code{copyright} and \code{tagline}.
All \code{header} arguments are character strings.
The \code{paper} argument list options are \code{textheight}, \code{linewidth}, \code{indent}, \code{first_page_number}, \code{page_numbers} and \code{fontsize}.
All \code{paper} arguments are numeric except \code{page_numbers}, which is logical.
}
\examples{
# not run
}
