% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes.R
\name{note-metadata}
\alias{note-metadata}
\alias{n_steps}
\alias{n_notes}
\alias{n_chords}
\alias{n_octaves}
\alias{octave_type}
\alias{accidental_type}
\alias{time_format}
\alias{is_space_time}
\alias{is_vector_time}
\alias{note_is_tick}
\alias{note_is_integer}
\alias{note_has_tick}
\alias{note_has_integer}
\title{Noteworthy string metadata}
\usage{
n_steps(notes)

n_notes(notes)

n_chords(notes)

n_octaves(notes)

octave_type(notes)

accidental_type(notes)

time_format(notes)

is_space_time(notes)

is_vector_time(notes)

note_is_tick(notes)

note_is_integer(notes)

note_has_tick(notes)

note_has_integer(notes)
}
\arguments{
\item{notes}{character, a noteworthy string, space-delimited or vector of
individual entries.}
}
\value{
varies by function
}
\description{
Inspect basic metadata for noteworthy strings.
}
\details{
These functions inspect the basic metadata of noteworthy strings.
For functions that perform basic checks on strings, see
\code{\link{note-checks}}.

The \code{n_*} functions give summary totals of the number of timesteps,
number of individual note (non-chord) timesteps, number of chord time
steps, and the number of distinct octaves present across timesteps.

Functions pertaining to type or format of a noteworthy string provide
information on how a particular string is defined, e.g. \code{time_format}.
Note that the result pertains to true \code{noteworthy}-class objects. If
inspecting a standard character string, the result pertains to
post-conversion to the \code{noteworthy} class and does not necessarily
reflect what is found in \code{notes} verbatim. See examples.
}
\examples{
x <- "e_2 a_, c#f#a#"
n_steps(x)
n_notes(x)
n_chords(x)
n_octaves(x)

# Type is mixed in \\code{x} but is inferred under default conversion rules.
# These check \\code{x} once validated and coerced to 'noteworthy' class.
octave_type(x)
accidental_type(x)
# The default is tick octaves and flats
as_noteworthy(x)

time_format(x)
is_space_time(x)
is_vector_time(x)
}
\seealso{
\code{\link{note-checks}}, \code{\link{note-summaries}},
\code{\link{note-coerce}}, \code{\link{valid-notes}}
}
