% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rest}
\alias{rest}
\title{Create rests}
\usage{
rest(x, n = 1)
}
\arguments{
\item{x}{integer, duration.}

\item{n}{integer, number of repetitions.}
}
\value{
a character string.
}
\description{
Create multiple rests efficiently with a simple wrapper around \code{rep} using the \code{times} argument.
}
\examples{
rest(c(1, 8), c(1, 4))
}
