% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{Create tablature}
\usage{
tab(score, file, key = "c", time = "4/4", tempo = "2 = 60",
  header = NULL, paper = NULL, endbar = TRUE, midi = TRUE,
  out_dir = ".", keep_ly = FALSE)
}
\arguments{
\item{score}{a score object.}

\item{file}{character, output file, should end in .pdf or .png.}

\item{key}{character, key signature, e.g., \code{c}, \code{b_}, \code{f#m}, etc.}

\item{time}{character, defaults to \code{"4/4"}.}

\item{tempo}{character, defaults to \code{"2 = 60"}.}

\item{header}{a named list of arguments passed to the header of the LilyPond file. See details.}

\item{paper}{a named list of arguments for the LilyPond file page layout. See details.}

\item{endbar}{character, the end bar.}

\item{midi}{logical, output midi file in addition to tablature.}

\item{out_dir}{character, output directory, defaults to working directory.}

\item{keep_ly}{logical, keep LilyPond file.}
}
\value{
nothing returned; a file is written.
}
\description{
Create sheet music/guitar tablature from a music score.
}
\details{
This function generates a pdf (optionally a png) of a music score using the LilyPond music engraving program.
Output format is inferred from \code{file} extension.
}
\examples{
# not run
}
