% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabr.R
\docType{package}
\name{tabr}
\alias{tabr}
\alias{tabr-package}
\title{tabr: Guitar tablature and sheet music engraving.}
\description{
The \code{tabr} package provides programmatic music notation and wraps around the open source music engraving program, Lilypond, for creating quality guitar tablature.
}
