% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{tabr_options}
\alias{tabr_options}
\title{Options}
\usage{
tabr_options(...)
}
\arguments{
\item{...}{a list of options.}
}
\value{
The function prints all set options if called with no arguments.
When setting options, nothing is returned.
}
\description{
Options for tabr package.
}
\details{
Currently only \code{lilypond}, \code{midi2ly} and \code{python} are used.
On Windows systems, if the system path for lilypond.exe, midi2ly.py and
python.exe are not stored in the system PATH environmental variable, they
must be provided by the user after loading the package.
}
\examples{
tabr_options()
lilypond_path <- "C:/Program Files (x86)/LilyPond/usr/bin/lilypond.exe"
tabr_options(lilypond = lilypond_path)
}
